Changes in service of making create_empty_s3_file take an s3_encrypt_key_id argument.
