+++ src/encoded/root.py
@@ -2,13 +2,14 @@ import uptime
 
 from collections import OrderedDict
 from dcicutils import lang_utils
-from dcicutils.s3_utils import s3Utils, HealthPageKey
+from dcicutils.s3_utils import HealthPageKey  # , s3Utils
 from dcicutils.env_utils import infer_foursight_url_from_env
 from pyramid.decorator import reify
 from pyramid.security import ALL_PERMISSIONS, Allow, Authenticated, Deny, Everyone
 from snovault import Root, calculated_property, root, COLLECTIONS, STORAGE
 from .appdefs import APP_VERSION_REGISTRY_KEY
 from .schema_formats import is_accession
+from .util import SettingsKey
 
 
 def includeme(config):
@@ -60,30 +61,6 @@ def uptime_info():
         return lang_utils.relative_time_string(uptime.uptime())
     except Exception:
         return "unavailable"
-
-
-class SettingsKey:
-    APPLICATION_BUCKET_PREFIX = 'application_bucket_prefix'
-    BLOB_BUCKET = 'blob_bucket'
-    EB_APP_VERSION = 'eb_app_version'
-    ELASTICSEARCH_SERVER = 'elasticsearch.server'
-    ENCODED_VERSION = 'encoded_version'
-    FILE_UPLOAD_BUCKET = 'file_upload_bucket'
-    FILE_WFOUT_BUCKET = 'file_wfout_bucket'
-    FOURSIGHT_BUCKET_PREFIX = 'foursight_bucket_prefix'
-    IDENTITY = 'identity'
-    INDEXER = 'indexer'
-    INDEXER_NAMESPACE = 'indexer.namespace'
-    INDEX_SERVER = 'index_server'
-    LOAD_TEST_DATA = 'load_test_data'
-    METADATA_BUNDLES_BUCKET = 'metadata_bundles_bucket'
-    S3_ENCRYPT_KEY_ID = 's3_encrypt_key_id'
-    SNOVAULT_VERSION = 'snovault_version'
-    SQLALCHEMY_URL = 'sqlalchemy.url'
-    SYSTEM_BUCKET = 'system_bucket'
-    TIBANNA_CWLS_BUCKET = 'tibanna_cwls_bucket'
-    TIBANNA_OUTPUT_BUCKET = 'tibanna_output_bucket'
-    UTILS_VERSION = 'utils_version'
 
 
 def health_check(config):
@@ -284,7 +261,6 @@ class CGAPRoot(Root):
         #     except KeyError:
         #         pass
         # return return_list
-
 
     @calculated_property(schema={
         "title": "Application version",
